+++ src/Queue.tpp
@@ -8,7 +8,7 @@
 #include "../include/Queue.h"
 
 template<typename T>
-Queue<T>::Queue(unsigned long long init_capacity) : size(0), capacity(init_capacity) {
+Queue<T>::Queue(unsigned int init_capacity) : size(0), capacity(init_capacity) {
     data = new T[init_capacity];
 }
 
@@ -16,7 +16,7 @@ template<typename T>
 Queue<T>::Queue(const Queue<T>& other) : size(other.size), capacity(other.capacity) {
     if (capacity > 0) {
         data = new T[capacity];
-        for (unsigned long long i = 0; i < size; i++)
+        for (unsigned int i = 0; i < size; i++)
             data[i] = other.data[i];
     }
 }
@@ -28,14 +28,14 @@ Queue<T>::~Queue() {
 
 template<typename T>
 void Queue<T>::print() {
-    for (unsigned long long i = 0llu; i < size; i++) {
-        printf("%llu ", data[i]);
+    for (unsigned int i = 0; i < size; i++) {
+        printf("%u ", data[i]);
     }
     printf("\n");
 }
 
 template<typename T>
-unsigned long long Queue<T>::getSize() const {
+unsigned int Queue<T>::getSize() const {
     return size;
 }
 
@@ -48,7 +48,7 @@ template<typename T>
 void Queue<T>::resize() {
     T* new_data = new T[capacity * RESIZE_SCALE];
 
-    for (unsigned long long i = 0llu; i < size; i++)
+    for (unsigned int i = 0; i < size; i++)
         new_data[i] = data[i];
 
     delete[] data;
@@ -75,8 +75,8 @@ T Queue<T>::dequeue() {
     if (size <= 0)
         throw std::out_of_range("queue is empty");
     T value = data[0];
-    for (unsigned long long i = 1llu; i < size; i++)
-        data[i-1llu] = data[i];
+    for (unsigned int i = 1; i < size; i++)
+        data[i-1] = data[i];
     size--;
     return value;
 }
@@ -88,7 +88,7 @@ T& Queue<T>::head() {
 
 template<typename T>
 T& Queue<T>::tail() {
-    return data[size-1llu];
+    return data[size-1];
 }
 
 #endif //GRAPHS_QUEUE_TPP
