+++ src/Vector.tpp
@@ -11,12 +11,12 @@
 #include "../include/Vector.h"
 
 template<typename T>
-Vector<T>::Vector() : size(0llu), capacity(INIT_CAPACITY) {
+Vector<T>::Vector() : size(0), capacity(INIT_CAPACITY) {
     data = new T[capacity];
 }
 
 template<typename T>
-Vector<T>::Vector(unsigned long long init_capacity) : size(0), capacity(init_capacity) {
+Vector<T>::Vector(unsigned int init_capacity) : size(0), capacity(init_capacity) {
     data = new T[capacity];
 }
 
@@ -24,7 +24,7 @@ template<typename T>
 Vector<T>::Vector(const Vector<T>& other) : size(other.size), capacity(other.capacity), data(nullptr) {
     if (capacity > 0) {
         data = new T[capacity];
-        for (unsigned long long i = 0; i < size; i++)
+        for (unsigned int i = 0; i < size; i++)
             data[i] = other.data[i];
     }
 }
@@ -35,15 +35,15 @@ Vector<T>::~Vector() {
 }
 
 template<typename T>
-unsigned long long Vector<T>::getSize() const {
+unsigned int Vector<T>::getSize() const {
     return this->size;
 }
 
 template<typename T>
-void Vector<T>::resize(unsigned long long new_capacity) {
+void Vector<T>::resize(unsigned int new_capacity) {
     T* new_data = new T[new_capacity];
 
-    for (unsigned long long i = 0llu; i < size; i++)
+    for (unsigned int i = 0; i < size; i++)
         new_data[i] = data[i];
     delete[] data;
 
@@ -60,32 +60,32 @@ void Vector<T>::push_back(const T& value) {
 
 template<typename T>
 void Vector<T>::print() const {
-    for (unsigned long long i = 0llu; i < size; i++)
-        printf("%llu ", data[i]);
+    for (unsigned int i = 0; i < size; i++)
+        printf("%u ", data[i]);
     printf("\n");
 }
 
 template<typename T>
 bool Vector<T>::isSorted() const {
-    for (unsigned long long i = 0llu; i < size-1llu; i++) {
-        if (data[i] < data[i+1llu])
+    for (unsigned int i = 0; i < size-1; i++) {
+        if (data[i] < data[i+1])
             return false;
     }
     return true;
 }
 
 template<typename T>
-void Vector<T>::sort(unsigned long long leftIter, unsigned long long rightIter) {
+void Vector<T>::sort(unsigned int leftIter, unsigned int rightIter) {
     if (leftIter >= rightIter || rightIter >= size || isSorted())
         return;
 
-    unsigned long long pivot = partition(leftIter, rightIter);
+    unsigned int pivot = partition(leftIter, rightIter);
     sort(leftIter, pivot);
-    sort(pivot+1llu, rightIter);
+    sort(pivot+1, rightIter);
 }
 
 template <typename T>
-unsigned long long Vector<T>::partition(unsigned long long leftIter, unsigned long long rightIter) {
+unsigned int Vector<T>::partition(unsigned int leftIter, unsigned int rightIter) {
     T pivot = data[leftIter];
 
     while (true) {
@@ -116,7 +116,7 @@ Vector<T>& Vector<T>::operator=(const Vector<T> &other) {
         data = nullptr;
         if (capacity > 0) {
             data = new T[capacity];
-            for (unsigned long long i = 0; i < size; ++i) {
+            for (unsigned int i = 0; i < size; ++i) {
                 data[i] = other.data[i];
             }
         }
@@ -125,14 +125,14 @@ Vector<T>& Vector<T>::operator=(const Vector<T> &other) {
 }
 
 template<typename T>
-T& Vector<T>::operator[](unsigned long long index) {
+T& Vector<T>::operator[](unsigned int index) {
     if (index >= size)
         throw std::out_of_range("index out of range");
     return data[index];
 }
 
 template<typename T>
-const T& Vector<T>::operator[](unsigned long long index) const {
+const T& Vector<T>::operator[](unsigned int index) const {
     if (index >= size)
         throw std::out_of_range("index out of range");
     return data[index];
